refactor: move user command backpressure to log stream module
